requirements.txt
about-time==4.2.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
grapheme==0.6.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.1.0
opencv-python==4.10.0.84
pillow==10.4.0
progress==1.6
pycparser==2.22
rich==13.7.1
sounddevice==0.5.0
soundfile==0.12.1
typing_extensions==4.12.2